+++ github/tests/ReplayData/Repository.testGetViews.txt
@@ -1,0 +1,11 @@
+https
+GET
+api.github.com
+None
+/repos/jkufro/PyGithub/traffic/views
+{'Authorization': 'Basic login_and_password_removed', 'User-Agent': 'PyGithub/Python'}
+None
+200
+[('Server', 'GitHub.com'), ('Date', 'Sun, 02 Dec 2018 16:01:47 GMT'), ('Content-Type', 'application/json; charset=utf-8'), ('Transfer-Encoding', 'chunked'), ('Status', '200 OK'), ('X-RateLimit-Limit', '5000'), ('X-RateLimit-Remaining', '4988'), ('X-RateLimit-Reset', '1543769966'), ('Cache-Control', 'private, max-age=60, s-maxage=60'), ('Vary', 'Accept, Authorization, Cookie, X-GitHub-OTP'), ('ETag', 'W/"c716db6b3f790a8d8abfa33a70b8ab9a"'), ('X-GitHub-Media-Type', 'github.v3; format=json'), ('Access-Control-Expose-Headers', 'ETag, Link, Location, Retry-After, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval, X-GitHub-Media-Type'), ('Access-Control-Allow-Origin', '*'), ('Strict-Transport-Security', 'max-age=31536000; includeSubdomains; preload'), ('X-Frame-Options', 'deny'), ('X-Content-Type-Options', 'nosniff'), ('X-XSS-Protection', '1; mode=block'), ('Referrer-Policy', 'origin-when-cross-origin, strict-origin-when-cross-origin'), ('Content-Security-Policy', "default-src 'none'"), ('Content-Encoding', 'gzip'), ('X-GitHub-Request-Id', 'F041:56B2:D5A72E:1E01B44:5C0401EB')]
+{"count":84,"uniques":4,"views":[{"timestamp":"2018-11-27T00:00:00Z","count":56,"uniques":4},{"timestamp":"2018-11-28T00:00:00Z","count":6,"uniques":2},{"timestamp":"2018-11-30T00:00:00Z","count":18,"uniques":1},{"timestamp":"2018-12-01T00:00:00Z","count":4,"uniques":1}]}
+
